avoid having multiple components defined in one file 🚨
