+++ src/app.stories.tsx
@@ -1,11 +1,9 @@
 import React from "react";
 
-import App, { AppHeader, AppFooter } from "app";
+import App from "app";
 
 export default {
   title: "Templates/App",
 };
 
 export const Home = () => <App />;
-export const Header = () => <AppHeader />;
-export const Footer = () => <AppFooter />;
