+++ src/components/layout/index.tsx
@@ -1,0 +1,2 @@
+export { default as FooterLayout } from "./footer";
+export { default as HeaderLayout } from "./header";
