+++ src/components/layout/layout.stories.tsx
@@ -1,0 +1,10 @@
+import React from "react";
+
+import { HeaderLayout, FooterLayout } from "./";
+
+export default {
+  title: "Organisms/Layout",
+};
+
+export const Header = () => <HeaderLayout />;
+export const Footer = () => <FooterLayout />;
